(cms): Move Undo, Redo, Export to ComposerHeader
