+++ apps/cms/src/components/ComposerHeader/ComposerHeader.module.css
@@ -54,3 +54,7 @@ svg.action-button {
   stroke-width: var(--editor-toolbar-icon-stroke);
   color: var(--color-menu-text-light);
 }
+
+button:disabled svg.action-button {
+  color: var(--color-menu-text-disabled);
+}
